+++ src/onegov.gazette/onegov/gazette/theme/gazette_theme.py
@@ -7,7 +7,7 @@ class GazetteTheme(BaseTheme):
 
     # don't touch this number, it's incremented using bumpversion, so every
     # release will automatically trigger a rebuild of the theme
-    version = '1.17.4'
+    version = '1.18.0'
 
     @property
     def post_imports(self):